Projects: Create rather than own project importers
The importers are only needed temporarily. This is now expressed by
having Projects transfer ownership of created ProjectImporter instances
using std::unique_ptr.

Moreover, TargetSetupPage held a QPointer to the project's importer,
indicating that it had to worry about the importer's lifetime. This is
resolved now by TargetSetupPage owning the importer via std::unique_ptr.
In order to really decouple the lifetimes of a ProjectImporter and its
Project, the project file path is passed as an argument rather than the
ProjectImporterCreator capturing a pointer to its Project instance.

Change-Id: Ia3766de97cd6873a24cb4a6b8637cf9e897605e7